Move validation code to overture-schema-validation
Moves validate command, type_analysis, error_formatting,
data_display to the new package. types.py retains only
ValidationErrorDict and ErrorLocation (ModelDict and UnionType
are in core).

output.py (rewrap) is deliberately duplicated -- both packages
keep a copy. A future CLI helper package will consolidate
shared utilities.

The validate command is now loaded by the CLI via entry point.
CLI package is stripped down to built-in commands (list-types,
json-schema).
