+++ packages/overture-schema-cli/src/overture/schema/cli/__init__.py
@@ -1,23 +1,5 @@
 """CLI subpackage for overture-schema."""
 
-from .commands import (
-    cli,
-    handle_generic_error,
-    handle_validation_error,
-    load_input,
-    perform_validation,
-)
-from .types import (
-    ErrorLocation,
-    ValidationErrorDict,
-)
+from .commands import cli
 
-__all__ = [
-    "cli",
-    "handle_generic_error",
-    "handle_validation_error",
-    "load_input",
-    "perform_validation",
-    "ErrorLocation",
-    "ValidationErrorDict",
-]
+__all__ = ["cli"]
