+++ packages/overture-schema-validation/src/overture/schema/validation/commands.py
@@ -1,0 +1,527 @@
+"""Validation command for overture-schema CLI."""
+
+import builtins
+import json
+import sys
+from collections import Counter, defaultdict
+from pathlib import Path
+from typing import cast
+
+import click
+import yaml
+from pydantic import BaseModel, TypeAdapter, ValidationError
+from rich.console import Console
+from yamlcore import CoreLoader  # type: ignore
+
+from overture.schema.core.discovery import resolve_types
+from overture.schema.core.union import UnionType
+
+from .error_formatting import (
+    format_validation_error,
+    format_validation_errors_verbose,
+    group_errors_by_discriminator,
+    select_most_likely_errors,
+)
+from .output import rewrap
+from .type_analysis import StructuralTuple, get_item_index, introspect_union
+from .types import ErrorLocation
+
+# Console instances for rich output
+stdout = Console(highlight=False)
+stderr = Console(highlight=False, file=sys.stderr)
+
+
+def _is_geojson_feature(data: dict) -> bool:
+    """Check if data is in GeoJSON Feature format."""
+    return data.get("type") == "Feature" and "properties" in data
+
+
+def validate_feature(data: dict, model_type: UnionType) -> BaseModel:
+    """Validate a single feature against the model type.
+
+    Args
+    ----
+        data: Feature data to validate (GeoJSON or flat format)
+        model_type: Union type for validation
+
+    Returns
+    -------
+        Validated model instance
+
+    Raises
+    ------
+        ValidationError: If validation fails
+    """
+    adapter = TypeAdapter(model_type)
+    if isinstance(data, dict) and _is_geojson_feature(data):
+        # Use validate_json to trigger the model's GeoJSON handling
+        return cast(BaseModel, adapter.validate_json(json.dumps(data)))
+    return cast(BaseModel, adapter.validate_python(data))
+
+
+def validate_features(data: list, model_type: UnionType) -> list[BaseModel]:
+    """Validate a list of features against the model type.
+
+    Args
+    ----
+        data: List of feature data to validate (GeoJSON or flat format)
+        model_type: Union type for validation
+
+    Returns
+    -------
+        List of validated model instances
+
+    Raises
+    ------
+        ValidationError: If validation fails
+    """
+    # Check if any items are GeoJSON features
+    has_geojson = any(
+        isinstance(item, dict) and _is_geojson_feature(item) for item in data
+    )
+
+    list_type = list[model_type]  # type: ignore[misc,valid-type]
+    adapter = TypeAdapter(list_type)
+
+    if has_geojson:
+        # Use validate_json to trigger the model's GeoJSON handling
+        return cast(list[BaseModel], adapter.validate_json(json.dumps(data)))
+    return cast(list[BaseModel], adapter.validate_python(data))
+
+
+def get_source_name(filename: Path) -> str:
+    """Get display name for input source.
+
+    Args
+    ----
+        filename: Path to input file or "-" for stdin
+
+    Returns
+    -------
+        Display name: "<stdin>" for stdin input, otherwise the filename
+    """
+    return "<stdin>" if str(filename) == "-" else str(filename)
+
+
+def load_input(filename: Path) -> tuple[dict | list, str]:
+    """Load and parse input from file or stdin.
+
+    Args
+    ----
+        filename: Path to input file, or "-" for stdin
+
+    Returns
+    -------
+        Tuple of (parsed_data, source_name)
+
+    Raises
+    ------
+        yaml.YAMLError: If input is invalid YAML/JSON
+        SystemExit: If filename doesn't exist or isn't a file
+    """
+    if str(filename) == "-":
+        # Read all stdin content
+        content = sys.stdin.read()
+
+        # Try to detect JSONL format (newline-delimited JSON)
+        # JSONL has multiple non-empty lines, each containing a complete JSON object
+        lines = [line.strip() for line in content.strip().split("\n") if line.strip()]
+
+        if len(lines) > 1:
+            # Attempt to parse as JSONL
+            try:
+                parsed_lines = [json.loads(line) for line in lines]
+                return parsed_lines, "<stdin>"
+            except json.JSONDecodeError:
+                # Not valid JSONL, fall through to YAML parser
+                pass
+
+        # Parse as single YAML/JSON document
+        import io
+
+        data = yaml.load(io.StringIO(content), Loader=CoreLoader)
+        return data, "<stdin>"
+
+    if not filename.is_file():
+        raise click.UsageError(f"'{filename}' is not a file.")
+
+    # Warn about unexpected file extensions
+    if filename.suffix not in {".json", ".yaml", ".yml", ".geojson"}:
+        click.echo(
+            f"Warning: File '{filename}' has unexpected extension. "
+            f"Expecting .json, .yaml, .yml, or .geojson",
+            err=True,
+        )
+
+    # Use YAML-1.2-compliant loader (YAML-1.2 dropped support for yes/no boolean values)
+    with filename.open("r", encoding="utf-8") as f:
+        data = yaml.load(f, Loader=CoreLoader)
+
+    return data, str(filename)
+
+
+def perform_validation(data: dict | list, model_type: UnionType) -> None:
+    """Validate data based on its structure.
+
+    Automatically detects and handles three input formats:
+    - Single feature (dict)
+    - List of features (list)
+    - GeoJSON FeatureCollection (dict with type="FeatureCollection")
+
+    Args
+    ----
+    data : dict | list
+        Parsed data to validate
+    model_type : UnionType
+        Union type for validation
+
+    Raises
+    ------
+    ValidationError
+        If validation fails
+    """
+    if isinstance(data, list):
+        # List of features
+        validate_features(data, model_type)
+    elif isinstance(data, dict) and data.get("type") == "FeatureCollection":
+        # GeoJSON FeatureCollection
+        validate_features(data["features"], model_type)
+    else:
+        # Single feature
+        validate_feature(data, model_type)
+
+
+def compute_collection_statistics(
+    item_types: dict[int, builtins.type[BaseModel] | None],
+    filtered_errors: list,
+) -> tuple[
+    int,
+    Counter[builtins.type[BaseModel] | None],
+    dict[builtins.type[BaseModel], set[int]],
+]:
+    """Compute validation statistics for heterogeneous collections.
+
+    Args
+    ----
+    item_types : dict[int, type[BaseModel] | None]
+        Mapping from item index to detected model type
+    filtered_errors : list
+        List of filtered validation errors
+
+    Returns
+    -------
+    tuple
+        Tuple of (items_without_errors, type_counts, items_with_errors_by_type)
+    """
+    # Compute statistics: group items by type
+    type_counts: Counter[builtins.type[BaseModel] | None] = Counter(item_types.values())
+
+    # Determine total number of items (max index + 1, or count from data)
+    max_index = max(item_types.keys()) if item_types else -1
+    total_items = max_index + 1
+
+    # Count items with errors per type
+    items_with_errors_by_type: dict[builtins.type[BaseModel], set[int]] = {}
+    for err in filtered_errors:
+        idx = get_item_index(err["loc"])
+        if idx is not None and idx in item_types:
+            model_type_cls = item_types[idx]
+            if model_type_cls is not None:
+                if model_type_cls not in items_with_errors_by_type:
+                    items_with_errors_by_type[model_type_cls] = set()
+                items_with_errors_by_type[model_type_cls].add(idx)
+
+    # Count items without any errors
+    items_without_errors = total_items - len(
+        {
+            idx
+            for idx in item_types.keys()
+            if any(get_item_index(err["loc"]) == idx for err in filtered_errors)
+        }
+    )
+
+    return items_without_errors, type_counts, items_with_errors_by_type
+
+
+def print_collection_statistics(
+    items_without_errors: int,
+    type_counts: Counter[builtins.type[BaseModel] | None],
+    items_with_errors_by_type: dict[builtins.type[BaseModel], set[int]],
+    stderr: Console,
+) -> None:
+    """Print validation statistics for heterogeneous collections.
+
+    Args
+    ----
+    items_without_errors : int
+        Count of items with no validation errors
+    type_counts : Counter[type[BaseModel] | None]
+        Counter of items by model type
+    items_with_errors_by_type : dict[type[BaseModel], set[int]]
+        Mapping from model type to set of item indices with errors
+    stderr : Console
+        Console for stderr output
+    """
+    stderr.print("  [dim]Collection statistics:[/dim]")
+
+    # Show items without errors first
+    # TODO: Once we switch to parse_features (instead of validate_features),
+    # we can include type information for items without errors by parsing
+    # the input and tracking which items validated successfully and their types.
+    # This would allow output like: "Building: 2 confirmed (no errors)"
+    if items_without_errors > 0:
+        stderr.print(
+            f"    • {items_without_errors} item{'s' if items_without_errors != 1 else ''} with no errors",
+            style="dim",
+        )
+
+    # Show per-type statistics
+    for model_type_cls, count in type_counts.most_common():
+        if model_type_cls is not None:
+            items_with_errors = len(
+                items_with_errors_by_type.get(model_type_cls, set())
+            )
+            valid_count = count - items_with_errors
+
+            if valid_count > 0:
+                stderr.print(
+                    f"    • {model_type_cls.__name__}: {valid_count} confirmed, {items_with_errors} with errors",
+                    style="dim",
+                )
+            else:
+                stderr.print(
+                    f"    • {model_type_cls.__name__} (probable): {items_with_errors} item{'s' if items_with_errors != 1 else ''} with errors",
+                    style="dim",
+                )
+    stderr.print()
+
+
+def handle_validation_error(
+    e: ValidationError,
+    model_type: UnionType,
+    stderr: Console,
+    original_data: dict | list | None = None,
+    show_fields: list[str] | None = None,
+) -> None:
+    """Handle and format validation errors with rich contextual information.
+
+    Groups errors by discriminator, selects most likely error groups, and provides
+    helpful diagnostics for heterogeneous collections and ambiguous types.
+
+    Args
+    ----
+    e : ValidationError
+        ValidationError from pydantic
+    model_type : UnionType
+        Union type used for validation
+    stderr : Console
+        Console for stderr output
+    original_data : dict | list | None
+        Original input data for error display
+    show_fields : list[str] | None
+        List of field names to display alongside errors
+    """
+    # Compute metadata once upfront
+    metadata = introspect_union(model_type)
+
+    # Create cache for structural tuple computation (optimizes systematic errors)
+    structural_cache: dict[ErrorLocation, StructuralTuple] = {}
+
+    # Group errors by discriminator path and select most likely group(s)
+    error_groups = group_errors_by_discriminator(e.errors(), metadata, structural_cache)
+    filtered_errors, is_tied, is_heterogeneous, item_types = select_most_likely_errors(
+        error_groups,
+        metadata=metadata,
+        all_errors=e.errors(),
+        structural_cache=structural_cache,
+    )
+
+    # Show heterogeneity warning if collection has mixed types
+    if is_heterogeneous:
+        stderr.print(
+            "  ⚠ Heterogeneous collection: Data contains multiple feature types.",
+            style="yellow",
+        )
+        stderr.print(
+            "    • Consider validating each type separately with --theme or --type",
+            style="dim",
+        )
+        stderr.print()
+
+        # Compute and display statistics if there are errors to report
+        if filtered_errors:
+            items_without_errors, type_counts, items_with_errors_by_type = (
+                compute_collection_statistics(item_types, filtered_errors)
+            )
+            print_collection_statistics(
+                items_without_errors, type_counts, items_with_errors_by_type, stderr
+            )
+
+    # Show tie indicator if multiple groups had same error count
+    elif is_tied:
+        stderr.print(
+            "  ⚠ Ambiguous: Data matches multiple types equally. Consider:",
+            style="yellow",
+        )
+        stderr.print(
+            "    • Specifying --theme or --type to narrow validation", style="dim"
+        )
+        stderr.print("    • Adding discriminator fields to clarify intent", style="dim")
+        stderr.print()
+
+    # Group errors by item
+
+    errors_by_item: dict[int | None, list] = defaultdict(list)
+    for error in filtered_errors:
+        item_idx = get_item_index(error["loc"])
+        errors_by_item[item_idx].append(error)
+
+    # Display errors grouped by item
+
+    for item_idx, item_errors in errors_by_item.items():
+        # Determine item type
+        error_item_type = None
+        if item_idx is not None and item_idx in item_types:
+            error_item_type = item_types.get(item_idx)
+
+        # Try verbose display first
+        displayed = format_validation_errors_verbose(
+            item_errors,
+            stderr,
+            metadata=metadata,
+            item_type=error_item_type,
+            structural_cache=structural_cache,
+            original_data=original_data,
+            item_index=item_idx,
+            show_fields=show_fields,
+        )
+
+        # Fall back to non-verbose format if verbose couldn't display
+        if not displayed:
+            for i, error in enumerate(item_errors):
+                format_validation_error(
+                    error,
+                    stderr,
+                    metadata=metadata,
+                    show_model_hint=(i == 0),
+                    item_type=error_item_type,
+                    show_item_type=is_heterogeneous,
+                    structural_cache=structural_cache,
+                    original_data=original_data,
+                    show_feature_data=False,
+                )
+
+
+def handle_generic_error(e: Exception, filename: Path, error_type: str) -> None:
+    """Handle generic errors during validation.
+
+    Args
+    ----
+    e : Exception
+        Exception that occurred
+    filename : Path
+        Input filename or "-" for stdin
+    error_type : str
+        Type of error for user-friendly message
+
+    Raises
+    ------
+    click.UsageError
+        Always, with formatted error message
+    """
+    source_name = get_source_name(filename)
+
+    if error_type == "yaml":
+        raise click.UsageError(f"'{source_name}' contains invalid input: {e}")
+    elif error_type == "value":
+        raise click.UsageError(str(e))
+    elif error_type == "key":
+        raise click.UsageError(f"Invalid data structure - missing key: {e}")
+    else:
+        raise click.UsageError(f"Error processing {source_name}: {e}")
+
+
+@click.command()
+@click.argument("filename", type=click.Path(path_type=Path), required=True)
+@click.option(
+    "--overture-types",
+    is_flag=True,
+    help="Validate against all official Overture types (excludes extensions)",
+)
+@click.option(
+    "--namespace",
+    help="Namespace to filter by (e.g., overture, annex)",
+)
+@click.option(
+    "--theme",
+    multiple=True,
+    help="Theme to validate against (shorthand for all types in theme)",
+)
+@click.option(
+    "--type",
+    "types",
+    multiple=True,
+    help="Specific type to validate against (e.g., building, segment)",
+)
+@click.option(
+    "--show-field",
+    "show_fields",
+    multiple=True,
+    help="Field to display alongside errors (e.g., id, version). Can be repeated.",
+)
+def validate(
+    filename: Path,
+    overture_types: bool,
+    namespace: str | None,
+    theme: tuple[str, ...],
+    types: tuple[str, ...],
+    show_fields: tuple[str, ...],
+) -> None:
+    r"""Validate Overture Maps data against schemas.
+
+    Read from FILENAME or stdin if FILENAME is '-'.
+    Supports JSON, YAML, and GeoJSON formats.
+
+    \b
+    Examples:
+      # Validate a file
+      $ overture-schema validate data.json
+    \b
+      # Validate from stdin
+      $ overture-schema validate - < data.json
+    \b
+      # Validate only buildings
+      $ overture-schema validate --theme buildings data.json
+    \b
+      # Validate specific type
+      $ overture-schema validate --type building data.json
+    \b
+      # Official Overture types only
+      $ overture-schema validate --overture-types data.json
+    """
+    # Resolve model type first (errors here are ValueErrors, not ValidationErrors)
+    try:
+        model_type = resolve_types(overture_types, namespace, theme, types)
+    except ValueError as e:
+        handle_generic_error(e, filename, "value")
+        return
+
+    # Load input (errors here are YAMLErrors or ValueErrors, not ValidationErrors)
+    try:
+        data, source_name = load_input(filename)
+    except yaml.YAMLError as e:
+        handle_generic_error(e, filename, "yaml")
+        return
+    except KeyError as e:
+        handle_generic_error(e, filename, "key")
+        return
+
+    # Perform validation (now model_type and data are guaranteed to be defined)
+    try:
+        perform_validation(data, model_type)
+        stdout.print(f"✓ Successfully validated {source_name}")
+    except ValidationError as e:
+        handle_validation_error(
+            e, model_type, stderr, original_data=data, show_fields=list(show_fields)
+        )
+        sys.exit(1)
